+++ inlang/source-code/ide-extension/src/main.ts
@@ -20,9 +20,13 @@ import { recommendationBannerView } from "./utilities/recommendation/recommendat
 import { telemetry } from "./services/telemetry/implementation.js"
 import { version } from "../package.json"
 import { statusBar } from "./utilities/settings/statusBar.js"
+import { initErrorMonitoring } from "./services/error-monitoring/implementation.js"
 
 // Entry Point
 export async function activate(context: vscode.ExtensionContext): Promise<void> {
+	// Sentry Error Handling
+	initErrorMonitoring()
+
 	try {
 		vscode.commands.executeCommand("setContext", "inlang:hasProjectInWorkspace", false)
 		vscode.commands.executeCommand("setContext", "inlang:showRecommendationBanner", false)
@@ -151,9 +155,4 @@ function handleInlangErrors() {
 	if (inlangErrors.length > 0) {
 		console.error("Inlang VSCode Extension errors:", inlangErrors)
 	}
-}
-
-// Helper Functions
-export function getActiveTextEditor(): vscode.TextEditor | undefined {
-	return vscode.window.activeTextEditor
 }