+++ inlang/source-code/website/src/pages/documentation/+onBeforeRender.tsx
@@ -1,4 +1,5 @@
 import fs from "node:fs/promises"
+import ecosystemTableOfContents from "../../../../../documentation/ecosystem/tableOfContents.json"
 import sdkTableOfContents from "../../../../../documentation/sdk/tableOfContents.json"
 import pluginTableOfContents from "../../../../../documentation/plugin/tableOfContents.json"
 import lintRuleTableOfContents from "../../../../../documentation/lint-rule/tableOfContents.json"
@@ -29,8 +30,16 @@ const redirectLinks = [
 		to: "/m/zu942ln6/app-inlang-badge",
 	},
 	{
-		from: "/documentation/sdk",
-		to: "/documentation",
+		from: "/documentation/build-app",
+		to: "/documentation/sdk/build-app",
+	},
+	{
+		from: "/documentation/publish-to-marketplace",
+		to: "/documentation/sdk/publish-to-marketplace",
+	},
+	{
+		from: "/documentation/publish-guide",
+		to: "/documentation/sdk/publish-guide",
 	},
 ]
 
@@ -43,7 +52,6 @@ export default async function onBeforeRender(pageContext: any) {
 		pageContext.urlPathname === "/documentation"
 			? ""
 			: pageContext.urlPathname.replace("/documentation/", "")
-
 	// Look for redirects
 	for (const redirectLink of redirectLinks) {
 		if (redirectLink.from === pageContext.urlPathname)
@@ -54,9 +62,23 @@ export default async function onBeforeRender(pageContext: any) {
 	}
 
 	if (renderedMarkdown[slug] === undefined) {
-		// get sdk documentation
+		// get ecosystem documentation
+		for (const categories of Object.entries(ecosystemTableOfContents)) {
+			const [, pages] = categories
+
+			for (const page of pages) {
+				const text = await fs.readFile(
+					new URL(`inlang/documentation/ecosystem/${page.path}`, repositoryRoot),
+					"utf-8"
+				)
+				const markdown = await convert(text)
+				renderedMarkdown[page.slug] = markdown
+			}
+		}
+		//get sdk documentation
 		for (const categories of Object.entries(sdkTableOfContents)) {
 			const [, pages] = categories
+
 			for (const page of pages) {
 				const text = await fs.readFile(
 					new URL(`inlang/documentation/sdk/${page.path}`, repositoryRoot),