+++ inlang/source-code/website/src/pages/documentation/getTableOfContents.ts
@@ -2,6 +2,7 @@ import { currentPageContext } from "#src/renderer/state.js"
 import sdkTableOfContents from "../../../../../documentation/sdk/tableOfContents.json"
 import pluginTableOfContents from "../../../../../documentation/plugin/tableOfContents.json"
 import lintRuleTableOfContents from "../../../../../documentation/lint-rule/tableOfContents.json"
+import ecosystemTableOfContents from "../../../../../documentation/ecosystem/tableOfContents.json"
 
 export const getTableOfContents = () => {
 	switch (currentPageContext.urlParsed.pathname.split("/")[2]) {
@@ -9,7 +10,9 @@ export const getTableOfContents = () => {
 			return pluginTableOfContents as typeof pluginTableOfContents
 		case "lint-rule":
 			return lintRuleTableOfContents as typeof lintRuleTableOfContents
-		default:
+		case "sdk":
 			return sdkTableOfContents as typeof sdkTableOfContents
+		default:
+			return ecosystemTableOfContents as typeof ecosystemTableOfContents
 	}
 }