+++ inlang/source-code/website/src/pages/index/custom_section/Personas/index.tsx
@@ -28,11 +28,11 @@ const Personas = () => {
 	]
 
 	return (
-		<div class="pt-12 md:pt-20 flex flex-col items-center">
+		<div id="personas" class="pt-16 flex flex-col items-center">
 			<p class="bg-background px-4 py-1.5 rounded-full text-sm font-medium w-fit border shadow border-surface-300">
 				{m.home_personas_tag()}
 			</p>
-			<h2 class="font-bold text-2xl md:text-4xl text-surface-900 text-center mt-4">
+			<h2 class="font-bold text-2xl md:text-4xl text-surface-900 text-center mt-6">
 				{m.home_personas_title()}
 			</h2>
 			<div class="w-full flex justify-center gap-2 mt-8 flex-wrap">